requirements.txt
-cipy==1.5.2
absl-py==0.11.0
addict==2.4.0
aiohttp==3.7.3
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
blinker==1.4
botorch==0.3.0
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configs==3.0.3
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
gpytorch==1.1.1
grpcio==1.31.0
h5py==3.2.1
idna==2.10
imageio==2.8.0
imageio-ffmpeg==0.4.8
importlib-metadata==1.7.0
imutils==0.5.3
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.1
Jinja2==2.11.2
joblib==0.15.1
jsonpatch==1.28
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
llvmlite==0.34.0
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
moviepy==1.0.3
multidict==4.7.6
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.0.3
numba==0.51.2
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
omegaconf==2.0.6
opencv-python==4.2.0.34
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.0.0
pip==20.0.2
prettytable==3.5.0
proglog==0.1.10
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.13.0
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydensecrf==1.0rc2
pygame==2.1.2
Pygments==2.6.1
PyJWT==2.0.1
pylint==2.5.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
pytesseract==0.2.4
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.1.0
seaborn==0.10.1
Send2Trash==1.5.0
setuptools==47.1.1.post20200604
Sift==5.0.2
six==1.16.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.6.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.6.3
toml==0.10.1
torch==1.5.0
torchfile==0.1.0
torchnet==0.0.5.1
torchsummary==1.5.1
torchvision==0.6.0
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
visdom==0.1.8.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wheel==0.34.2
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
xlrd==1.2.0
yarl==1.5.1
zipp==3.1.0
